Added feature to generate database + schema name as extensions / annotations.
