--- TEAM/ClassExtensionMethods.cs
@@ -1,96 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Win32;
-
-namespace TEAM
-{
-    /// <summary>
-    /// Extension methods
-    /// </summary>
-    public static class ExtensionMethod
-    {
-        public static DataTable ToDataTable<T>(this IList<T> data)
-        {
-            PropertyDescriptorCollection props =
-            TypeDescriptor.GetProperties(typeof(T));
-            DataTable table = new DataTable();
-            for (int i = 0; i < props.Count; i++)
-            {
-                PropertyDescriptor prop = props[i];
-                table.Columns.Add(prop.Name, prop.PropertyType);
-            }
-            object[] values = new object[props.Count];
-            foreach (T item in data)
-            {
-                for (int i = 0; i < values.Length; i++)
-                {
-                    values[i] = props[i].GetValue(item);
-                }
-                table.Rows.Add(values);
-            }
-            return table;
-        }
-
-        private static string CleanifyBrowserPath(string p)
-        {
-            string[] url = p.Split('"');
-            string clean = url[1];
-            return clean;
-        }
-
-        public static string GetDefaultBrowserPath()
-        {
-            string urlAssociation = @"Software\Microsoft\Windows\Shell\Associations\UrlAssociations\http";
-            string browserPathKey = @"$BROWSER$\shell\open\command";
-
-            Microsoft.Win32.RegistryKey userChoiceKey = null;
-            string browserPath = "";
-
-            try
-            {
-                //Read default browser path from userChoiceLKey
-                userChoiceKey = Microsoft.Win32.Registry.CurrentUser.OpenSubKey(urlAssociation + @"\UserChoice", false);
-
-                //If user choice was not found, try machine default
-                if (userChoiceKey == null)
-                {
-                    //Read default browser path from Win XP registry key
-                    var browserKey = Registry.ClassesRoot.OpenSubKey(@"HTTP\shell\open\command", false);
-
-                    //If browser path wasn’t found, try Win Vista (and newer) registry key
-                    if (browserKey == null)
-                    {
-                        browserKey =
-                        Microsoft.Win32.Registry.CurrentUser.OpenSubKey(
-                        urlAssociation, false);
-                    }
-                    var path = CleanifyBrowserPath(browserKey.GetValue(null) as string);
-                    browserKey.Close();
-                    return path;
-                }
-                else
-                {
-                    // user defined browser choice was found
-                    string progId = (userChoiceKey.GetValue("ProgId").ToString());
-                    userChoiceKey.Close();
-
-                    // now look up the path of the executable
-                    string concreteBrowserKey = browserPathKey.Replace("$BROWSER$", progId);
-                    var kp = Registry.ClassesRoot.OpenSubKey(concreteBrowserKey, false);
-                    browserPath = CleanifyBrowserPath(kp.GetValue(null) as string);
-                    kp.Close();
-                    return browserPath;
-                }
-            }
-            catch (Exception)
-            {
-                return "";
-            }
-        }
-    }
-}
