+++ TEAM/TEAM.csproj
@@ -227,10 +227,9 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClassLoadPatternHandling.cs" />
-    <Compile Include="ClassUtility.cs" />
-    <Compile Include="ClassExtensionMethods.cs" />
-    <Compile Include="ClassObjectModel.cs" />
+    <Compile Include="LoadPatternHandling.cs" />
+    <Compile Include="Utility.cs" />
+    <Compile Include="TeamObjectModel.cs" />
     <Compile Include="CustomTabPageEnvironment.cs">
       <SubType>Component</SubType>
     </Compile>
@@ -240,14 +239,14 @@
     <Compile Include="CustomDataGridViewAttribute.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="ClassJsonHandling.cs" />
+    <Compile Include="JsonHandling.cs" />
     <Compile Include="CustomDataGridViewPhysicalModel.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="CustomTimedTextBox.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="ClassLocalTeamEnvironmentConfiguration.cs" />
+    <Compile Include="LocalTeamEnvironmentConfiguration.cs" />
     <Compile Include="Form_Base.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -308,7 +307,8 @@
     <Compile Include="Form_Validation.Designer.cs">
       <DependentUpon>Form_Validation.cs</DependentUpon>
     </Compile>
-    <Compile Include="ClassMetadataValidation.cs" />
+    <Compile Include="MetadataValidation.cs" />
+    <Compile Include="JsonOutputHandling.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CustomDataGridViewTable.cs">
